// src/libs/api/schema/core-services.d.ts
export interface components {
    schemas: {
// ...
        AddressType: "Main" | "Billing" | "Shipping" | "Other";
// ...
        ContactLegalType: "Individual" | "Partnership" | "Limited Partnership" | "Limited Liability Partnership" | "Limited Liability Company" | "Public Limited Company" | "Private Limited Company" | "Corporation" | "Nonprofit Organization" | "Cooperative" | "Trust" | "Government Agency";
// ...
        IndustryType: "Agriculture, Forestry, and Fishing" | "Mining and Quarrying" | "Manufacturing" | "Utilities (Electricity, Gas, and Water Supply)" | "Water Supply, Waste Management, and Remediation" | "Construction" | "Wholesale and Retail Trade" | "Transportation and Storage" | "Accommodation and Food Service Activities" | "Information and Communication" | "Financial and Insurance Activities" | "Real Estate Activities" | "Professional, Scientific, and Technical Services" | "Administrative and Support Services" | "Public Administration and Defense" | "Education" | "Human Health and Social Work Activities" | "Arts, Entertainment, and Recreation" | "Other Service Activities" | "Activities of Households as Employers" | "Activities of Extraterritorial Organizations";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}